Rank exact matches higher.
