+++ internal/server/v3/nodesearch/golden/node_search_test.go
@@ -33,6 +33,7 @@ const (
 )
 
 func TestV3NodeSearch(t *testing.T) {
+	test.EnableSpannerGraph = true
 	// TODO: Remove check once enabled.
 	if !test.EnableSpannerGraph {
 		return
